Add files via upload
The purpose of this study was to investigate the importance of monitoring Marine protected areas and to  identify tools and techniques that are appropriate for monitoring in marine protected areas. The importance of monitoring  marine protected areas is an increasing concern for stakeholders of Marine protected areas. Using a cross-sectional analysis, this study analyzed research articles from 2000 to 2018 and measured these results against a similar study conducted in 2022. Remote sensing technology was found to play the greatest role in facilitating the monitoring of marine protected areas within developed and developing nations. Additionally, this study has discovered that Omic technology is one of the newest bio technologies for monitoring the biodiversity of marine protected areas.   The study suggests that the monitoring of marine protected areas is a continuous process and it permits management of Marine protected areas to identify and evaluate the status of threats and address issues ,such as the reputation of conservation targets and the execution of conservation actions. Furthermore, the monitoring of marine protected areas allows stakeholders to appraise the effectiveness of management interventions. The study definitively answers the question regarding what are  the importance of monitoring protected areas and the tools and techniques used for monitoring marine protected areas. However, further studies are necessary to identify the challenges of monitoring marine protected areas and develop remedies for ameliorating those confrontations.